refactor: Moving hooks to apiHooks
